requirements.txt
numpy
openai
pandas
Requests
streamlit
streamlit_option_menu
xgboost
scikit-learn
matplotlib
seaborn
streamlit-lottie
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain
langchain-google-genai
python-dotenv
pyspark
mlflow
loguru
feedparser
